+++ extensions/fotmob/src/favorite.tsx
@@ -1,0 +1,5 @@
+import SearchView from "./views/search/SearchView";
+
+export default function Command() {
+  return <SearchView />;
+}