+++ extensions/music/src/favorite-and-add-to-library.tsx
@@ -6,6 +6,6 @@ import { handleTaskEitherError } from "./util/utils";
 
 export default pipe(
   music.currentTrack.favorite,
-  TE.chain(() => pipe(music.currentTrack.addToLibrary)),
+  TE.chainW(() => pipe(music.currentTrack.addToLibrary)),
   handleTaskEitherError("Failed to save/favorite the track", "Favorited & added to library")
 );