+++ extensions/notion/src/utils/notion/database/index.ts
@@ -11,6 +11,19 @@ import { DatabaseProperty, DatabasePropertyOption } from "./property";
 
 export type { DatabaseProperty, DatabasePropertyOption };
 
+export async function fetchDatabase(pageId: string) {
+  try {
+    await authorize();
+    const page = await notion.databases.retrieve({
+      database_id: pageId,
+    });
+
+    return pageMapper(page);
+  } catch (err) {
+    return handleError(err, "Failed to fetch database", undefined);
+  }
+}
+
 export async function fetchDatabases() {
   try {
     await authorize();
@@ -145,7 +158,7 @@ export async function createDatabasePage(values: Form.Values) {
     };
 
     if (content) {
-      // TODO: why do I need to cast this?
+      // casting because converting from the `Block` type in martian to the `BlockObjectRequest` type in notion
       arg.children = markdownToBlocks(content) as BlockObjectRequest[];
     }
 
@@ -154,6 +167,7 @@ export async function createDatabasePage(values: Form.Values) {
       if (!type) return;
       const propId = formId.match(new RegExp("(?<=property::" + type + "::).*", "g"))?.[0];
       const value = values[formId];
+      if (value == "_select_null_") return;
       if (!propId || !value) return;
 
       const formatted = formatDatabaseProperty(type, value);