+++ extensions/pins/src/lib/Pins.ts
@@ -5,11 +5,13 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-09-04 17:37:42
- * Last modified  : 2023-09-04 17:37:42
+ * Last modified  : 2023-11-01 00:43:50
  */
 
 import { useCachedState } from "@raycast/utils";
 import {
+  Alert,
+  Application,
   Clipboard,
   Keyboard,
   LaunchType,
@@ -22,14 +24,15 @@ import {
   showToast,
 } from "@raycast/api";
 import { useEffect, useState } from "react";
-import { StorageKey } from "./constants";
-import { ExtensionPreferences, getStorage, runCommand, runCommandInTerminal, setStorage } from "./utils";
+import { SORT_FN, StorageKey, SORT_STRATEGY } from "./constants";
+import { getStorage, runCommand, runCommandInTerminal, setStorage } from "./utils";
+import { ExtensionPreferences } from "./preferences";
 import * as fs from "fs";
 import * as os from "os";
 import { Placeholders } from "./placeholders";
 import path from "path";
 import { LocalDataObject } from "./LocalData";
-import { Group, SortStrategy } from "./Groups";
+import { Group } from "./Groups";
 
 /**
  * A pin object.
@@ -176,7 +179,7 @@ export const usePins = () => {
  * @param pin The pin to open.
  * @param preferences The extension preferences object.
  */
-export const openPin = async (pin: Pin, preferences: { preferredBrowser: string }, context?: LocalDataObject) => {
+export const openPin = async (pin: Pin, preferences: { preferredBrowser: Application }, context?: LocalDataObject) => {
   const startDate = new Date();
 
   try {
@@ -263,8 +266,24 @@ export const openPin = async (pin: Pin, preferences: { preferredBrowser: string 
     () => {
       null;
     },
-    false
+    false,
   );
+};
+
+/**
+ * Gets the next available pin ID.
+ */
+export const getNextPinID = async () => {
+  // Get the stored pins
+  const storedPins = await getStorage(StorageKey.LOCAL_PINS);
+
+  // Get the next available group ID
+  let newID = (await getStorage(StorageKey.NEXT_PIN_ID))[0] || 1;
+  while (storedPins.some((pin: Group) => pin.id == newID)) {
+    newID++;
+  }
+  setStorage(StorageKey.NEXT_PIN_ID, [newID + 1]);
+  return newID;
 };
 
 /**
@@ -289,7 +308,7 @@ export const createNewPin = async (
   execInBackground: boolean | undefined,
   fragment: boolean | undefined,
   shortcut: Keyboard.Shortcut | undefined,
-  iconColor: string | undefined
+  iconColor: string | undefined,
 ) => {
   // Get the stored pins
   const storedPins = await getStorage(StorageKey.LOCAL_PINS);
@@ -355,7 +374,7 @@ export const modifyPin = async (
   averageExecutionTime: number | undefined,
   pop: () => void,
   setPins: React.Dispatch<React.SetStateAction<Pin[]>>,
-  notify = true
+  notify = true,
 ) => {
   const storedPins = await getStorage(StorageKey.LOCAL_PINS);
   const newData: Pin[] = storedPins.map((oldPin: Pin) => {
@@ -424,8 +443,20 @@ export const modifyPin = async (
  * @param pin The pin to delete.
  * @param setPins The function to update the list of pins.
  */
-export const deletePin = async (pin: Pin, setPins: React.Dispatch<React.SetStateAction<Pin[]>>) => {
-  if (await confirmAlert({ title: "Are you sure?" })) {
+export const deletePin = async (
+  pin: Pin,
+  setPins: React.Dispatch<React.SetStateAction<Pin[]>>,
+  showAlert = true,
+  displayToast = true,
+) => {
+  if (
+    !showAlert ||
+    (await confirmAlert({
+      title: `Delete Pin '${pin.name}'`,
+      message: "Are you sure?",
+      primaryAction: { title: "Delete", style: Alert.ActionStyle.Destructive },
+    }))
+  ) {
     const storedPins = await getStorage(StorageKey.LOCAL_PINS);
 
     const filteredPins = storedPins.filter((oldPin: Pin) => {
@@ -434,7 +465,7 @@ export const deletePin = async (pin: Pin, setPins: React.Dispatch<React.SetState
 
     setPins(filteredPins);
     await setStorage(StorageKey.LOCAL_PINS, filteredPins);
-    await showToast({ title: `Removed pin!` });
+    if (displayToast) await showToast({ title: `Removed pin!` });
   }
 };
 
@@ -455,10 +486,20 @@ export const getPreviousPin = async (): Promise<Pin | undefined> => {
  * @param groups The list of groups to sort by.
  * @returns The sorted list of pins.
  */
-export const sortPins = (pins: Pin[], groups: Group[], sortMethod?: SortStrategy) => {
+export const sortPins = (
+  pins: Pin[],
+  groups: Group[],
+  sortMethod?: keyof typeof SORT_STRATEGY,
+  sortFn?: (a: unknown, b: unknown) => number,
+) => {
   const preferences = getPreferenceValues<ExtensionPreferences>();
   return [...pins].sort((p1, p2) => {
     const group = groups.find((group) => group.name == p1.group);
+    if (sortFn) {
+      // Use custom sort function if provided
+      return sortFn(p1, p2);
+    }
+
     if (
       sortMethod == "alphabetical" ||
       (sortMethod == undefined &&
@@ -500,12 +541,7 @@ export const sortPins = (pins: Pin[], groups: Group[], sortMethod?: SortStrategy
  * @returns The {@link Pin} that was most recently opened, or undefined if no pins have been created yet.
  */
 export const getLastOpenedPin = (pins: Pin[]) => {
-  const sortedPins = [...pins].sort((p1, p2) => {
-    return (
-      (p2.lastOpened ? new Date(p2.lastOpened) : new Date(0)).getTime() -
-      (p1.lastOpened ? new Date(p1.lastOpened) : new Date(0)).getTime()
-    );
-  });
+  const sortedPins = sortPins(pins, [], undefined, SORT_FN.LAST_OPENED);
   if (sortedPins.length > 0) {
     return sortedPins[0];
   }
@@ -518,12 +554,7 @@ export const getLastOpenedPin = (pins: Pin[]) => {
  * @returns The {@link Pin} that was most recently created, or undefined if no pins have been created yet.
  */
 export const getNewestPin = (pins: Pin[]) => {
-  const sortedPins = [...pins].sort((p1, p2) => {
-    return (
-      (p2.dateCreated ? new Date(p2.dateCreated) : new Date(0)).getTime() -
-      (p1.dateCreated ? new Date(p1.dateCreated) : new Date(0)).getTime()
-    );
-  });
+  const sortedPins = sortPins(pins, [], undefined, SORT_FN.NEWEST);
   if (sortedPins.length > 0) {
     return sortedPins[0];
   }
@@ -536,12 +567,7 @@ export const getNewestPin = (pins: Pin[]) => {
  * @returns The {@link Pin} that was created first, or undefined if no pins have been created yet.
  */
 export const getOldestPin = (pins: Pin[]) => {
-  const sortedPins = [...pins].sort((p1, p2) => {
-    return (
-      (p1.dateCreated ? new Date(p1.dateCreated) : new Date(0)).getTime() -
-      (p2.dateCreated ? new Date(p2.dateCreated) : new Date(0)).getTime()
-    );
-  });
+  const sortedPins = sortPins(pins, [], undefined, SORT_FN.OLDEST);
   if (sortedPins.length > 0) {
     return sortedPins[0];
   }
@@ -566,7 +592,7 @@ export const getTotalPinExecutions = (pins: Pin[]) => {
 export const calculatePinFrequencyPercentile = (pin: Pin, pins: Pin[]) => {
   const pinsSortedByFrequency = pins.sort((a, b) => (a.timesOpened || 0) - (b.timesOpened || 0));
   const pinIndex = pinsSortedByFrequency.findIndex(
-    (p) => p.id == pin.id || (p.timesOpened || 0) >= (pin.timesOpened || 0)
+    (p) => p.id == pin.id || (p.timesOpened || 0) >= (pin.timesOpened || 0),
   );
   return Math.round((pinIndex / pinsSortedByFrequency.length) * 100);
 };
@@ -580,7 +606,7 @@ export const calculatePinFrequencyPercentile = (pin: Pin, pins: Pin[]) => {
 export const calculatePinExecutionTimePercentile = (pin: Pin, pins: Pin[]) => {
   const pinsSortedByExecutionTime = pins.sort((a, b) => (a.averageExecutionTime || 0) - (b.averageExecutionTime || 0));
   const pinIndex = pinsSortedByExecutionTime.findIndex(
-    (p) => p.id == pin.id || (p.averageExecutionTime || 0) >= (pin.averageExecutionTime || 0)
+    (p) => p.id == pin.id || (p.averageExecutionTime || 0) >= (pin.averageExecutionTime || 0),
   );
   return Math.round((pinIndex / pinsSortedByExecutionTime.length) * 100);
 };
@@ -659,7 +685,6 @@ export const getPinStatistics = (pin: Pin, pins: Pin[], format: "string" | "obje
     : "Never";
 
   const percentOfAllExecutions = `${Math.round(((pin.timesOpened || 0) / getTotalPinExecutions(pins)) * 100)}%`;
-
   const averageExecutionTime = pin.averageExecutionTime ? `${pin.averageExecutionTime / 1000} seconds` : "N/A";
 
   const placeholdersUsed = Object.entries(Placeholders.allPlaceholders).filter(([regex]) => {
@@ -701,7 +726,9 @@ export const getPinStatistics = (pin: Pin, pins: Pin[], format: "string" | "obje
       : 0
   }`;
 
-  const percentOfAllExecutionsText = `${percentOfAllExecutions} of All Pin Executions`;
+  const percentOfAllExecutionsText = `${
+    percentOfAllExecutions == "NaN%" ? "0%" : percentOfAllExecutions
+  } of All Pin Executions`;
 
   const executionTimePercentile = calculatePinExecutionTimePercentile(pin, pins);
   const averageExecutionTimeText = `Average Execution Time: ${averageExecutionTime}${